+++ components/examples/temperature_color_display.py
@@ -5,6 +5,9 @@ import CodyNick
 
 cody = CodyNick.CN()
 
+alarm_enabled = True
+was_clicked = False
+
 
 while True:
     temperature = CodyNick.Temperature_Sensor.read(cody)
@@ -22,5 +25,23 @@ while True:
     for led in range(16):
         CodyNick.RGB_Matrix.set(cody, led, color)
 
+    is_clicked = CodyNick.Joystick.click(cody, "CJP")
+
+    if is_clicked and not was_clicked:
+        alarm_enabled = not alarm_enabled
+
+        if alarm_enabled:
+            print("Alarm enabled")
+        else:
+            print("Alarm inhibited")
+
+    was_clicked = is_clicked
+
+    if temperature > 30 and alarm_enabled:
+        CodyNick.CJP_Sound_Maker.play_until_done(cody, "C6", 120)
+        CodyNick.CJP_Sound_Maker.play_until_done(cody, "E6", 120)
+        CodyNick.CJP_Sound_Maker.play_until_done(cody, "G6", 120)
+        CodyNick.CJP_Sound_Maker.play_until_done(cody, "E6", 120)
+
     print("Temperature:", temperature, "C")
-    time.sleep(1)
+    time.sleep(0.3)